// minijira-backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Priority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]


export const Status = {
  backlog: 'backlog',
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done'
} as const

export type Status = (typeof Status)[keyof typeof Status]